pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "guidance-credibility"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30",
    "pydantic>=2.0",
    "httpx>=0.27",
    "numpy>=1.26",
    "pandas>=2.0",
    "yfinance>=0.2",
    "rich>=13.0",
    "matplotlib>=3.8",
    "python-dotenv>=1.0",
]

[project.scripts]
guidance-credibility = "guidance_credibility.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
